requirements.txt
telethon
better-automation==1.1.0
aiohttp-socks
better-proxy
python-dotenv
loguru
